refactor: nmr-cli
