+++ app/scripts/nmr-cli/Dockerfile
@@ -19,7 +19,7 @@ RUN npm install
 COPY . ./
 
 #install the nmr-cli as a global package 
-# for example, nmr-cli -u https://cheminfo.github.io/bruker-data-test/data/zipped/aspirin-1h.zip
+# for example, nmr-cli parse-spectra -u https://cheminfo.github.io/bruker-data-test/data/zipped/aspirin-1h.zip
 RUN npm install . -g
